refactor(humn): group runtime internals
